+++ R/function_PlotAnnualRegime.R
@@ -49,14 +49,16 @@
 #' @examples
 #' # Source data, HYPE basin output with a number of result variables
 #' te1 <- ReadBasinOutput(filename = system.file("demo_model",
-#' "results", "0003587.txt", package = "HYPEtools"))
+#'                                               "results", "0003587.txt",
+#'                                               package = "HYPEtools"))
 #' # Daily discharge regime, computed and observed,
 #' # hydrologigical year from October, aggregated to weekly means
-#' te2 <- AnnualRegime(te1[, c("DATE", "COUT", "ROUT")], ts.in = "day",
-#' ts.out = "week", start.mon = 10)
+#' te2 <- AnnualRegime(te1[, c("DATE", "COUT", "ROUT")],
+#'                     ts.in = "day",
+#'                     ts.out = "week", start.mon = 10)
 #' PlotAnnualRegime(x = te2)
 #' PlotAnnualRegime(x = te2, line = "median", band = "p05p95",
-#' add.legend = TRUE, col = c("red", "blue"))
+#'   add.legend = TRUE, col = c("red", "blue"))
 #' 
 #' @importFrom graphics par grid abline axis axis.POSIXct box polygon lines legend
 #' @importFrom stats na.omit