+++ NEWS.md
@@ -4,7 +4,7 @@
 *Development Version*
 
 #### Highlights
-
+* Add basic MergeObs() function
 
 ## HYPEtools 1.4.1 (2023-06-30)
 *Enhancements and bug fixes*